requirements.txt
# PDF Parsing
pdfplumber==0.11.4

# LangChain Framework
langchain==0.3.21
langchain-core==0.3.45
langchain-community==0.3.18
langchain-text-splitters==0.3.7
langchain-huggingface>=0.1.0

# Embeddings & Vector Store
sentence-transformers==2.6.1
chromadb==0.4.24

# LLM API
openai>=1.12.0
langchain-openai>=0.0.8

# Environment Configuration
python-dotenv==1.0.1
